stocks-crawler/requirements.txt
avro-to-python==0.3.3
certifi==2022.12.7
charset-normalizer==2.1.1
Click==7.0
distlib==0.3.7
dnspython==2.2.1
Faker==16.8.1
fastavro==1.8.4
filelock==3.12.4
idna==3.4
Jinja2==2.10.3
MarkupSafe==2.1.3
numpy==1.24.2
pandas==1.5.3
pandas-datareader==0.10.0
platformdirs==3.11.0
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.31
six==1.16.0
urllib3==1.26.13
virtualenv==20.24.5
yfinance==0.2.37
python-dotenv==1.0.1
